refactor(agent): 迁移并行执行器到 workflow/parallel 子包并以 NodeExecutor 接口打破循环
